Split out LanguageTag in its own file. TODO: tests!
